requirements.txt
# Janus BOLA Detection Tool - Requirements
# Professional Platform Edition with Advanced Modules

# === Core ===
requests>=2.28.0

# === Proxy ===
mitmproxy>=10.0.0

# === Storage ===
# Redis is optional - falls back to JSON file storage
redis>=4.0.0

# === CLI Interface ===
typer>=0.9.0
rich>=13.0.0

# === Web Interface ===
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# === Advanced Modules (Phase 3) ===
deepdiff>=6.0.0     # Smart Diff - noise-aware comparison
jinja2>=3.1.0       # HTML report generation

# === Live Intelligence (Phase 4) ===
aiohttp>=3.8.0      # Async HTTP for CVE lookup
pyyaml>=6.0.0       # OpenAPI YAML parsing

# === Target App ===
flask>=2.0.0

# === Utilities ===
colorama>=0.4.6
